Refactor FWFormViewController
